move current stash content list composables into a single-location variant, to be separated from multi-location
